Fix FreeBSD build
Under FreeBSD getline prototype is not provided by
default due to compatibility problems.
So first check in CMake with check_function_exists to
make sure that getline provided by the system.
Then try to compile small code to test if declaration
is also available. If not declare function in code.

Closes #3217.