vinyl: on space alter update key definitions in indexes
Vinyl alter can change comparators, for example, by removal of a
space format and making some fields optional. This alter is
allowed on a non-empty space even, because optionality of a
field is not persisted. But index key definitions can not be
updated in place, since they are used in multiple threads (vinyl
workers). So it is necessary to get new key definitions from a
new index definition, and leave old ones to be deleted by
existing worker tasks, when they will be finished.

Index is not the only place storing key definition pointers:
there are also mems, ranges, cache - all these structures must be
updated too.

Closes #3229